split out files and clean up code
