+++ Domain.Sql/Domain.Sql.csproj
@@ -110,6 +110,7 @@
     <Compile Include="EventHandlerProgress.cs" />
     <Compile Include="EventHandlerProgressCalculator.cs" />
     <Compile Include="EventStoreDatabaseInitializer{T}.cs" />
+    <Compile Include="Migrations\IDbMigrator.cs" />
     <Compile Include="Migrations\201407120005564_v0_8_2.cs" />
     <Compile Include="Migrations\201407120005564_v0_8_2.Designer.cs">
       <DependentUpon>201407120005564_v0_8_2.cs</DependentUpon>
@@ -142,7 +143,7 @@
     <Compile Include="EventBusExtensions.cs" />
     <Compile Include="ExclusiveEventStoreCatchupQuery.cs" />
     <Compile Include="EventContractResolver.cs" />
-    <Compile Include="DbContextExtensions.cs" />
+    <Compile Include="DatabaseExtensions.cs" />
     <Compile Include="CopyEventStoreIfNotExists.cs" />
     <Compile Include="ReservationServiceDbContext.cs" />
     <Compile Include="ExceptionExtensions.cs" />
@@ -169,6 +170,7 @@
     <Compile Include="CommandScheduler\ScheduledCommand.cs" />
     <Compile Include="CommandScheduler\SqlCommandSchedulerPipelineInitializer{T}.cs" />
     <Compile Include="CommandSchedulerResolver.cs" />
+    <Compile Include="Migrations\ScriptBasedDbMigrator.cs" />
     <Compile Include="Sensors.cs" />
     <Compile Include="SqlReservationService.cs" />
     <Compile Include="StorableEvent.cs" />
